pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
paho-mqtt = "2.1.0"
pydantic = "2.10.6"
pydantic-settings = "2.8.1"
pyyaml = "6.0.2"
python-slugify = "8.0.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "3.8.0"

[[tool.mypy.overrides]]
module = [
    "slugify",
    "yaml",
]
ignore_missing_imports = true
